+++ R/col_factor.R
@@ -6,12 +6,15 @@
 #' @param ... unquoted column names for modification
 #' @param .all if `TRUE` then any column with a `values` attribute or  where 
 #' the `make_factor` field evaluates to `TRUE` will be added as a factor
+#' @param .missing a label to use assign to missing values `NA` when making 
+#' the factor; keep this `NULL` (the default) to let missing values be handled
+#' naturally by `factor()`
 #' @param .suffix used to make the column name for the factors
-#' @param values a vector of values to convert to a factor
-#' @param x a ycol object
 #' 
 #' @details
-#' Note that `.suffix` can be chosen using option `ys.fct.suffix`. 
+#' Note that `.suffix` can be chosen using option `ys.fct.suffix`. When the 
+#' factor is made by [base::factor()], the `exclude` argument is forced to 
+#' `character(0)` so that nothing is excluded.
 #' 
 #' @examples
 #' 
@@ -31,10 +34,11 @@
 #' @md
 #' @export
 ys_add_factors <- function(.data, .spec, ... , 
-                           .all = TRUE, 
+                           .all = TRUE, .missing = NULL,
                            .suffix = getOption("ys.fct.suffix","_f")) {
   
   assert_that(inherits(.spec, "yspec"))
+  assert_that(is.null(.missing) || is.character(.missing))
   
   fct_ok <- map_lgl(.spec, ~ isTRUE(.x[["make_factor"]]))
   
@@ -49,7 +53,12 @@ ys_add_factors <- function(.data, .spec, ... ,
   
   for(v in vars) {
     newcol <- paste0(v, .suffix)
-    .data[[newcol]] <- ys_make_factor(.data[[v]],.spec[[v]],strict=!fct_ok[[v]])
+    .data[[newcol]] <- ys_make_factor(
+      .data[[v]],
+      .spec[[v]],
+      strict=!fct_ok[[v]], 
+      .missing = .missing
+    )
   }
   .data
 }
@@ -58,10 +67,15 @@ ys_add_factors <- function(.data, .spec, ... ,
 #' @export
 yspec_add_factors <- ys_add_factors
 
-#' @param strict if `FALSE`, then an factor will be returned for any `values` type
+#' @param values a vector of values to convert to a factor
+#' @param x a ycol object
+#' @param strict if `FALSE`, then a factor will be returned for any `values` type
 #' @rdname ys_add_factors
 #' @export
-ys_make_factor <- function(values,x,strict=TRUE) {
+ys_make_factor <- function(values, x, strict = TRUE, .missing = NULL) {
+  if(is.factor(values)) {
+    return(values)
+  }
   if(is.null(x[["values"]])) {
     if(!strict) return(factor(values))
     stop("column: ", x[["col"]], " - values field is not found", call. = FALSE)
@@ -74,7 +88,12 @@ ys_make_factor <- function(values,x,strict=TRUE) {
   } else {
     decode <- x[["decode"]]  
   }
-  factor(values, levels = x[["values"]], labels = decode)
+  if(!is.null(.missing) && anyNA(values)) {
+    values[is.na(values)] <- .missing
+    x[["values"]] <- c(x[["values"]], .missing)
+    decode <- c(decode, .missing)
+  }
+  factor(values, levels = x[["values"]], labels = decode, exclude = character(0))
 }
 
 #' @rdname ys_add_factors