+++ tests/testthat/test-namespace.R
@@ -5,6 +5,7 @@ library(dplyr)
 context("test-namespace")
 works <- system.file("spec", "test", "namespace-works.yaml", package = "yspec")
 
+
 test_that("load a yaml file and parse namespaces", {
   spec <- ys_load(works)
   expect_message(ys_namespace(spec), msg = "namespaces:")
@@ -45,3 +46,15 @@ test_that("alternate decode in namespace", {
   expect_error(ys_load(file), msg = "decode is not the correct length")
 })
 
+test_that("multiple namespaces are handled properly", {
+  x <- yspec:::test_spec_test("namespace-multiple.yml")
+  plot <- ys_namespace(x, "plot")
+  expect_identical(x$STUDY$short, "short base")
+  expect_identical(plot$STUDY$short, "short plot")
+  lc <- ys_namespace(x, "lc")
+  uc <- ys_namespace(x, "uc")
+  expect_identical(x$STUDY$decode, c("Phase 1", "Phase 2", "Phase 3"))
+  expect_identical(plot$STUDY$decode, c("Phase 1", "Phase 2", "Phase 3"))
+  expect_identical(lc$STUDY$decode, c("phase 1", "phase 2", "phase 3"))
+  expect_identical(uc$STUDY$decode, c("PHASE 1", "PHASE 2", "PHASE 3"))
+})